+++ pybis/src/python/pybis/sample.py
@@ -4,6 +4,8 @@ from .openbis_object import OpenBisObject
 from .definitions import openbis_definitions
 from .utils import VERBOSE
 
+sample_definitions = openbis_definitions('Sample')
+
 class Sample(OpenBisObject):
     """ A Sample is one of the most commonly used objects in openBIS.
     """