+++ pybis/src/python/tests/test_person.py
@@ -15,16 +15,10 @@ def test_create_person(openbis_instance):
     person.save()
     person_exists = openbis_instance.get_person(userId=person_name)
     assert person_exists is not None
-
-    person.delete("test on {}".format(timestamp))
-
-    with pytest.raises(ValueError):
-        person_not_exists = openbis_instance.get_person(userId=person_name)
-        assert person_not_exists is None
-
-
-def test_role_assignments(openbis_instance, role_assignment_person):
-    person = openbis_instance.get_person(userId=role_assignment_person)
+    
+    # test role assignments
+    roles = person.get_roles()
+    assert len(roles) == 0
 
     for role in person.get_roles():
         role.delete('test')
@@ -32,18 +26,27 @@ def test_role_assignments(openbis_instance, role_assignment_person):
     roles_exist = person.get_roles()
     assert len(roles_exist) == 0
 
-    person.assign_role('ADMIN')
+    person.assign_role('INSTANCE_OBSERVER')
     roles_exist = person.get_roles()
     assert len(roles_exist) == 1
 
-    person.assign_role(role='ADMIN', space='DEFAULT')
+    person.assign_role(role='SPACE_OBSERVER', space='DEFAULT')
     roles_exist = person.get_roles()
     assert len(roles_exist) == 2
 
-    person.revoke_role(role='ADMIN')
+    person.revoke_role(role='INSTANCE_OBSERVER')
     roles_exist = person.get_roles()
     assert len(roles_exist) == 1
 
-    person.revoke_role(role='ADMIN', space='DEFAULT')
+    person.revoke_role(role='SPACE_OBSERVER', space='DEFAULT')
     roles_exist = person.get_roles()
     assert len(roles_exist) == 0
+
+
+
+    person.delete("test on {}".format(timestamp))
+
+    with pytest.raises(ValueError):
+        person_not_exists = openbis_instance.get_person(userId=person_name)
+        assert person_not_exists is None
+